баг на visiting_screen, текстовые константы в один файл
