+++ lib/mocks.dart
@@ -22,8 +22,8 @@ final List mocks = <Sight>[
         'Военно-исторический музей и развлекательный комплекс в Москве в 5-м Котельническом переулке рядом со станцией метро «Таганская». Основан в 2006 году. Расположен в подземном бункере площадью более 7000 м² на глубине 65 метров на территории бывшего засекреченного военного объекта СССР - Запасного командного пункта дальней авиации.',
     imgPreview:
         'https://rests.afisha.ru/uploads/images/1/22/122eb468a03f4baea1d3b7cf002d9744.jpg',
-    // favorites: WhereShowCard.visited,
-    // date: '12 окт. 2020',
+    favorites: WhereShowCard.visited,
+    date: '12 окт. 2020',
   ),
   Sight(
     name: 'Нереальное место',
