+++ lib/ui/screen/sight_card.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
 import 'package:places/domain/sight.dart';
-import 'package:places/ui/screen/visiting_screen_constant.dart';
+import 'package:places/ui/screen/res/strings.dart';
 
 /// в зависимости от места показа карточки - Список поиска, в Избранном
 /// (запланировано, посещено) показываем разную информацию на карточке
