--- lib/ui/screen/sight_details_constant.dart
@@ -1,3 +1,0 @@
-const buttonTitleToSchedule = 'Запланировать';
-const buttonTitleAddToFavourites = 'В Избранное';
-const buttonTitleBuildRoute = 'ПОСТРОИТЬ МАРШРУТ';
