--- lib/ui/screen/sight_list_screen_constant.dart
@@ -1,1 +1,0 @@
-const appBarTitle = 'Список\nинтересных мест';
